requirements.txt
astunparse==1.6.3
autopep8==1.5.7
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
Pillow==8.2.0
pycodestyle==2.7.0
six==1.15.0
tensorflow-estimator==2.5.0
toml==0.10.2
Werkzeug==2.0.1
